+++ examples/vue/lib/app.js
@@ -1,5 +1,3 @@
-
-
 const Koa = require('koa')
 const Router = require('koa-router')
 const app = new Koa()