+++ lib/pkg.ts
@@ -1,4 +1,4 @@
 export default {
-  version: '1.1.0-beta.15',
+  version: '1.1.0',
   name: 'aliyun-pds-js-sdk',
 }